refactor: update imports to use source-transformer-builder for consistency
