+++ packages/adapters/src/runners/source-transformer-runner.ts
@@ -1,5 +1,5 @@
 import type { Cache, CacheOptions } from "@mojis/internal-utils";
-import type { AnySourceTransformer } from "../builders/version-builder/types";
+import type { AnySourceTransformer } from "../builders/source-transformer-builder/types";
 import type { AdapterContext, SourceAdapterType } from "../global-types";
 import { fetchCache } from "@mojis/internal-utils";
 import { genericParse } from "@mojis/parsers";
