+++ packages/adapters/test/runners/composite-runner.test.ts
@@ -4,7 +4,7 @@ import { HttpResponse, mockFetch } from "#msw-utils";
 import { type } from "arktype";
 import { describe, expect, expectTypeOf, it } from "vitest";
 import { createCompositeHandlerBuilder } from "../../src/builders/composite-builder/builder";
-import { createSourceTransformerBuilder } from "../../src/builders/version-builder/builder";
+import { createSourceTransformerBuilder } from "../../src/builders/source-transformer-builder/builder";
 import { createFakeSourceAdapter, setupAdapterTest } from "../__utils";
 
 describe("run composite handler", () => {
